+++ 02_BackEnd/02_Exercicios/Lista02/src/ex09/entities/CommonEmployee.java
@@ -11,6 +11,6 @@ public class CommonEmployee extends Employee {
     }
 
     public double calculateSalary(int workedDays){
-        return this.getBaseSalary() * (workedDays/30) + this.getIncrease() * workedDays;
+        return this.getBaseSalary() * ((double)workedDays/30) + this.getIncrease() * workedDays;
     }
 }
